feat: move custom CSS command to core-commands package
